This commit has a few things.
First, and most importantly, the numcharacters.cfg file is able to control the number of characters created. Right now, don't put in more than 7, because there are only 7 paths in the database. But the ability is there....I just need to create a lot more paths. Cool? Cool.

Second, I started to create an optionMenu, which will have the ability  to change how many characters the user wants easily, without having to go to the config file. It's a work in progress, but won't break anything. Yup. have fun. -CM